+++ modules/appeals_api/app/controllers/appeals_api/v0/appeals_controller.rb
@@ -6,15 +6,15 @@ module AppealsApi
   module V0
     class AppealsController < ApplicationController
       skip_before_action(:authenticate)
-      skip_before_action :log_request, only: [:healthcheck]
 
       def index
+        log_request
         appeals_response = Appeals::Service.new.get_appeals(
           target_veteran,
           'Consumer' => consumer,
           'VA-User' => requesting_va_user
         )
-        log_response(appeals_log_attributes(appeals_response))
+        log_response(appeals_response)
         render(
           json: appeals_response.body
         )
@@ -26,11 +26,36 @@ module AppealsApi
 
       private
 
-      def appeals_log_attributes(appeals_response)
-        {
-          'first_appeal_id' => appeals_response.body['data'][0]['id'],
-          'appeal_count' => appeals_response.body['data'].length
-        }
+      def log_request
+        hashed_ssn = Digest::SHA2.hexdigest ssn
+        Rails.logger.info('Caseflow Request',
+                          'va_user' => requesting_va_user,
+                          'lookup_identifier' => hashed_ssn)
+      end
+
+      def log_response(appeals_response)
+        first_appeal_id = appeals_response.body['data'][0]['id']
+        count = appeals_response.body['data'].length
+        Rails.logger.info('Caseflow Response',
+                          'va_user' => requesting_va_user,
+                          'first_appeal_id' => first_appeal_id,
+                          'appeal_count' => count)
+      end
+
+      def consumer
+        request.headers['X-Consumer-Username']
+      end
+
+      def ssn
+        ssn = request.headers['X-VA-SSN']
+        raise Common::Exceptions::ParameterMissing, 'X-VA-SSN' unless ssn
+        ssn
+      end
+
+      def requesting_va_user
+        va_user = request.headers['X-VA-User']
+        raise Common::Exceptions::ParameterMissing, 'X-VA-User' unless va_user
+        va_user
       end
 
       def target_veteran
