--- modules/appeals_api/app/models/appeals_api/evss_claim.rb
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-module AppealsApi
-  class EVSSClaim
-    include Virtus.model
-    include ActiveModel::Serialization
-
-    attribute :evss_id, Integer
-    attribute :data, Hash
-    attribute :list_data, Hash
-
-    def requested_decision
-      false
-    end
-
-    def updated_at
-      nil
-    end
-  end
-end
