+++ modules/appeals_api/config/routes.rb
@@ -6,16 +6,11 @@ AppealsApi::Engine.routes.draw do
   namespace :v0, defaults: { format: 'json' } do
     resources :appeals, only: [:index]
     get 'healthcheck', to: 'appeals#healthcheck'
-    resources :claims, only: %i[index show]
   end
 
   namespace :docs do
     namespace :v0 do
-      # DEPRECATED DOC LINK (To be removed)
-      get 'api', to: 'api#appeals'
-      # -------------------
-      get 'appeals', to: 'api#appeals'
-      get 'claims', to: 'api#claims'
+      resources :api, only: [:index]
     end
   end
 end
