+++ modules/appeals_api/spec/requests/api_docs_request_spec.rb
@@ -3,17 +3,9 @@
 require 'rails_helper'
 
 RSpec.describe 'Appeals Status Documentation Endpoint', type: :request do
-  describe '#get /docs/v0/appeals' do
+  describe '#get /docs/v0/api' do
     it 'should return Open API Spec v3 JSON' do
-      get '/services/appeals/docs/v0/appeals'
-      expect(response).to have_http_status(:ok)
-      JSON.parse(response.body)
-    end
-  end
-
-  describe '#get /docs/v0/claims' do
-    it 'should return Open API Spec v3 JSON' do
-      get '/services/appeals/docs/v0/claims'
+      get '/services/appeals/docs/v0/api'
       expect(response).to have_http_status(:ok)
       JSON.parse(response.body)
     end
